# api/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise Fleet Management Platform API
  version: 1.0.0-locked
  description: >-
    Locked API contract for the Enterprise Fleet Management Platform.
    This document is generated from docs/architecture/00-locked-decisions.md and
    the PostgreSQL schema in db/schema. The frontend team binds the externally
    supplied UI designs to these endpoints (decision A3.6). No UI prescription
    is contained here.

    Conventions (decisions D1, D4, D7, C5.1):
      * Base path /api/v1. JSON only.
      * All timestamps are RFC 3339 in UTC (stored as timestamptz, rendered EAT client-side per A2.3).
      * Cursor pagination via `cursor` / `limit` query params; cursors are opaque.
      * Errors use RFC 7807 problem+json with an `error_code` extension member.
      * Every state-changing request MUST carry `Idempotency-Key: <uuid>` (C5.1).
  contact:
    name: Fleet Platform Architecture
servers:
  - url: https://api.fleet.internal/api/v1
    description: Production (AWS af-south-1, decision A1.10)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    idempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key
...
  schemas:
    Problem:
      type: object
      description: RFC 7807 problem+json.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, example: https://docs.fleet.internal/problems/clockout_pending }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uuid }
        error_code:
          type: string
          description: Stable machine-readable code (see Error Codes).
          example: CLOCKOUT_PENDING
        field_errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
...
    AcceptInviteRequest:
      type: object
      description: Unauthenticated. The single-use token is the credential; a password is set on first use.
      required: [token, password]
      properties:
        token: { type: string, format: uuid }
        password: { type: string, minLength: 12, maxLength: 200 }
        full_name: { type: string, maxLength: 200 }
...
    SignupRequest:
      type: object
      description: >
        Unauthenticated self-service onboarding. Creates the company (app.tenants, TRIAL/BASIC) and
        its first member, who is granted a TENANT-SCOPED ADMIN role — the company super-admin. That
        account sees its whole company and may invite further ADMIN / FLEET_MANAGER users, but it is
        never a cross-tenant SYSTEM_ADMIN. The tenant slug is derived from company_name server-side.
      required: [company_name, email, password]
      properties:
        company_name: { type: string, minLength: 1, maxLength: 200 }
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 12, maxLength: 200 }
        full_name: { type: string, minLength: 1, maxLength: 200 }
...
    LoginResponse:
      type: object
      description: Authenticated session (mirrors the mobile client's trusted response body).
      required: [token_type, access_token, refresh_token, user_id, roles, permissions, locale]
      properties:
        token_type: { type: string, example: Bearer }
        access_token: { type: string }
        access_token_expires_at: { type: string, format: date-time }
        refresh_token: { type: string }
        refresh_token_expires_at: { type: string, format: date-time }
        session_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        roles: { type: array, items: { type: string } }
        permissions: { type: array, items: { type: string } }
        locale: { type: string, enum: [en, sw] }
        tenant_id: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Malformed request (RFC7807).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing / expired token.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
    NotFound:
      description: Resource not found.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Conflict:
      description: Optimistic / state conflict.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate by email + password (+ MFA if enrolled).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                mfa_code: { type: string, description: Required when the user has MFA enrolled (A3.7). }
                device_id_hash: { type: string, description: Required for driver PIN offline revocation path (B12). }
      responses:
        '200':
          description: Tokens issued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  refresh_token: { type: string }
                  mfa_required: { type: boolean }
                  roles: { type: array, items: { type: string } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Account suspended (B13).
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  /auth/mfa/enroll:
    post:
      tags: [Auth]
      summary: Begin TOTP enrolment (Admin / Fleet Manager only, A3.7).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        '200':
          description: Provisioning URI + recovery codes (recovery codes shown once).
          content:
            application/json:
              schema:
                type: object
                properties:
                  provisioning_uri: { type: string, format: uri }
                  secret_encrypted_preview: { type: string }
                  recovery_codes: { type: array, items: { type: string } }
...
  /auth/signup:
    post:
      tags: [Tenancy]
      summary: Create a company and its first tenant-scoped ADMIN (self-service).
      description: >
        Unauthenticated and idempotent. Creates app.tenants (TRIAL/BASIC) with a slug derived from
        company_name, creates the user, and grants ADMIN inside that new tenant. The response is the
        standard session body plus the created tenant_id, so the client is signed in immediately.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignupRequest' }
      responses:
        '201':
          description: Company created; session issued for its first ADMIN.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/LoginResponse'
                  - type: object
                    properties:
                      tenant_id: { type: string, format: uuid }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/accept-invite:
    post:
      tags: [Tenancy]
      summary: Accept an invitation, set a password, join the inviter's tenant.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AcceptInviteRequest' }
      responses:
        '200':
          description: Session issued for the newly created member.
          content: { application/json: { schema: { $ref: '#/components/schemas/LoginResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...